feat(http-traffic-record): add authorization to list http traffic records
This commit introduces an authorization check in the `ListHttpTrafficRecordsUseCase`.

The use case now ensures that the user requesting the list of HTTP traffic records is the owner of the specified traffic source. If the user is not the owner, a 'Forbidden' error is returned.

This prevents a user from being able to view traffic data from a source they do not own.